+++ test/integration/commander-spellbook.test.ts
@@ -3,13 +3,6 @@
 import spellbook = require("../../src/");
 
 describe("Commander Spellbook", () => {
-  afterAll(() => {
-    // if we don't do this after the tests complete,
-    // the test hangs as we're holding onto a reference
-    // to the promise of the reesolved API response
-    spellbook.resetCache();
-  });
-
   it("looks up all combos", async () => {
     const combos = await spellbook.search("");
 
@@ -60,5 +53,42 @@ describe("Commander Spellbook", () => {
 
       expect(hasOffColorCombo).toBeFalsy();
     });
+  });
+
+  it("looks up specific prequisite in combos", async () => {
+    const combos = await spellbook.search("prerequisites:permanents");
+
+    expect(combos.length).toBeGreaterThan(0);
+    const hasWordPermanentsInComboPreq = combos.every((combo) => {
+      return combo.prerequisites.find(
+        (prereq) => prereq.toLowerCase().indexOf("permanents") > -1
+      );
+    });
+
+    expect(hasWordPermanentsInComboPreq).toBeTruthy();
+  });
+
+  it("looks up specific step in combos", async () => {
+    const combos = await spellbook.search("steps:Tap");
+
+    expect(combos.length).toBeGreaterThan(0);
+    const hasWordTapInSteps = combos.every((combo) => {
+      return combo.steps.find((step) => step.toLowerCase().indexOf("tap") > -1);
+    });
+
+    expect(hasWordTapInSteps).toBeTruthy();
+  });
+
+  it("looks up specific result in combos", async () => {
+    const combos = await spellbook.search("result:Infinite");
+
+    expect(combos.length).toBeGreaterThan(0);
+    const hasWordInfiniteInResult = combos.every((combo) => {
+      return combo.result.find(
+        (res) => res.toLowerCase().indexOf("infinite") > -1
+      );
+    });
+
+    expect(hasWordInfiniteInResult).toBeTruthy();
   });
 });