+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+# unreleased
+
+- Require manual cache busting
+
 # 0.1.1
+
 - Fix issue where `getScryfallImageUrl` would not return proper version
 
 # 0.1.0
 
-Initial release
+- Initial release